Restructure the package layout
A rather massiv change to the project structure. We now have the
binaries located in a bin/ folder and dedicated subdirectories for the
individual compartments of the lpjguesstools package. Helper files
currently only used by lgt_createinput are still within the scope of
lgt_createinput but should move to a dedicated tools folder later so
other scripts can access them.