Add country arg to MapContainer
Adding the optional country arg to Map or MapContainer triggers if a
country should be highlighted in the drawmap stage.

The country identifier is the admin0 3-letter short. Country can be
passed as a string, list or dict.

Depending on the arg, one or multiple countries are highlighted and
optionally filled with the provided color.

Examples:
country='CHL': Chile is highlighted, no fill
country=['CHL','ARG']: Chile and Argentina are highlighted, no fill
country={'CHL':{'fill': 'red'}, 'ARG': {'fill': None}}: CHL and ARG are
highlighted, CHL is filled red